+++ examples/api-gateway/Main.hs
@@ -2,13 +2,15 @@ module Main where
 
 import AWS.Lambda.APIGatewayInputEvent
 import AWS.Lambda.APIGatewayOutputEvent hiding (body)
+import AWS.Lambda.Encoding
 import AWS.Lambda.Handler
 import Control.Lens
 import Control.Monad.Logger
 import Control.Monad.IO.Class
+import Control.Monad.Except
 
 main :: IO ()
 main = runStderrLoggingT $ handler echo
 
-echo :: (MonadIO m, MonadLogger m) => APIGatewayInputEvent -> m APIGatewayOutputEvent
+echo :: (LambdaEncode e, MonadIO m, MonadLogger m, MonadError e m) => APIGatewayInputEvent -> m APIGatewayOutputEvent
 echo ie = pure $ apiGatewayOutputEvent $ ie ^. body
