+++ examples/dhall/Main.hs
@@ -9,6 +9,7 @@ import Dhall
 import AWS.Lambda.Encoding
 import Control.Monad.Logger
 import Control.Monad.IO.Class
+import Control.Monad.Except
 
 data User = User {
   name :: Text, 
@@ -20,5 +21,5 @@ data User = User {
 main :: IO ()
 main = runStderrLoggingT $ handler echo
 
-echo :: (MonadIO m, MonadLogger m) => User -> m User
+echo :: (LambdaEncode e, MonadIO m, MonadLogger m, MonadError e m) => User -> m User
 echo = pure
